+++ app/page.tsx
@@ -1,7 +1,7 @@
-import WhyNugenIT from "@/src/components/chooseUs";
-import TechnologyShowcase from "@/src/components/courses";
-import HeroSection from "@/src/components/heroSection";
-import StudentTestimonials from "@/src/components/testimonial";
+import WhyNugenIT from "@/src/components/homePage/chooseUs";
+import TechnologyShowcase from "@/src/components/homePage/courses";
+import HeroSection from "@/src/components/homePage/heroSection";
+import StudentTestimonials from "@/src/components/homePage/testimonialSection";
 import Footer from "@/src/layout/Footer";
 import Header from "@/src/layout/Header";
